Search for intersection at multiple higher orders
There was previously a problem with the "_compare_operation" method
in "inter[section]" mode -- it would stop after looking at only one
higher order (smaller cells).  Now in this mode it returns a list
of (order, cells) pairs.  If it finds an intersection at a lower
(larger cell) or the same order it returns a list containing only
the current cell, since that covers the whole area being searched.
Otherwise it builds a list of entries found at all smaller levels.